requirements.txt
langchain
langchain-openai
python-dotenv
fastapi
uvicorn[standard]
lark-oapi
